DAOS-11391 control,bio: Add MD-on-SSD role assignment to server config file
Allow NVMe SSD roles for MD-on-SSD to be explicitly set per-bdev-tier in
the server config file. The settings are data, index and wal. Settings
are communicated to the engine in the SPDK JSON config and exposed in
the bdev name once the SPDK subsystems have been initialized in the
engine's BIO module.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>