DAOS-10301 test: pass dmg command line with mmap()
It seems possible that the buffer where dmg command
line is being constructed, to be passed to system() call, can
be in a page that had been registered for RDMA, which can
cause a SEGV in fork()'ed child upon access for parsing.

So, to ensure allocating buffer in a brand new page, using mmap()
to avoid this.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>