DAOS-11757 common: Add DBTREE_CLASS_IFV for fixed size value
To avoid extra, unnecessary allocations in btree records when the values
are a fixed size, inline the user defined record.
We can apply this to both in-memory trees and durable format trees
because vea_format will run in the context of pool create and newly
created pools do not need to maintain backwards compatibility.  There
is no upgrade path from older pools for vea but this is a limitation
we can live with.

Modifies the compatibility code slightly to simplify setting of
compatibility flags.

Remove -x from run_test.sh which was recently added accidentally.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>