DAOS-11664 agent: Add config option to ignore interfaces
In some environments clients may be connected to multiple networks
that support the server's fabric provider, but some of them should
not or cannot be used for DAOS. This new config option allows unwanted
fabric interfaces to be excluded by name, rather than defining a full
fabric_ifaces configuration. Excluded interfaces will not be suggested
to clients.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>