DAOS-10138 chk: Schedule PS reconf before DONE
Prevent automatic PS reconfigurations in the CR "check" mode, in order
to avoid racing with the chk module's actions. Schedule PS
reconfigurations when the chk module is ready before moving a pool to
the DONE phase.

Fix a segfault due to the use of an invalid svc variable when destroying
a stopped rsvc replica:

  pool_svc_locate_cb (id=0x18, path=0x7fa947c6d2e0) at
    src/pool/srv_pool.c:953
  ds_rsvc_stop (class=DS_RSVC_CLASS_POOL, id=0x7fa96c5f0860,
    caller_term=18446744073709551615, destroy=<optimized out>) at
    src/rsvc/srv.c:1009
  ds_rsvc_stop_handler (rpc=0x7fa96c5f0660) at src/rsvc/srv.c:1437
  crt_handle_rpc (arg=0x7fa96c5f0660) at src/cart/crt_rpc.c:1654

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true