DAOS-11755 control: Allow daos_server --help without privilege
Also allow version and network scan commands that don't need the
helper to be running.

Move the call to pbin.CheckHelper from main() to parseOpts() just
before the call to Execute().  This allows the --help option to work
as that is handled by the option parsing logic.  Further, skip the
check if the command is version or network scan, as these don't need
privilege to run.

Required-githooks: true
Signed-off-by: Ted Anderson <nonzero@google.com>