DAOS-11724 test: pool/svc.py - Update the number of ranks to stop
In pool/svc.py, we stop 0, 1, or 2 ranks based on the expected
number of PS replicas. However, there's a limit for the number
of ranks we can stop, which are:

Number of PS replicas - Number of ranks we can stop
1 - 0
3 - 1
5 - 2

We were stopping 2 ranks when the number of PS replicas was 3.
Update the test so that we only stop 1 rank; and stop 2 ranks
only when the number of PS replicas is 5.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: test_pool_svc
Required-githooks: true
Signed-off-by: Makito Kano <makito.kano@intel.com>