DAOS-11145 csum: Scrubbing Upgrade Fix
When checksum scrubbing landed, it did not add the new properties
to the pool upgrade path. This patch fixes it so that a pool that was
created with 2.2 is able to be upgraded to a version with scrubbing
included.

Also did some minor renaming to make scrubbing RDB key consistent.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>