DAOS-5921 md: bypass space checks in pool evict rdb tx update (#10414)
Before this change if pool rdb metadata is exhausted, and if the pool has no open connections, a pool destroy operation fails with -DER_NOSPACE. This is because the evict handles step only performs an RDB tx update on the "connectable" metadata property. Updates are not marked as a critical operation (so far, only RDB tx delete and destroy operations have that special designation).

With this change, certain RDB TX updates (currently only this one) can be declared "critical" by invoking the new function rdb_tx_update_critical() rather than rdb_tx_update(). This results in bypassing the SCM free space checks when committing the TX, and ultimately results in successful operation.

Also with this change the functional test infrastructure allows a test to customize, when destroying a pool, if its containers will be destroyed first. The test_metadata_fillup test is changed to keep its containers (whose creates have filled the pool metadata). So the test in tearDown exercises pool destroy under metadata full conditions.

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>
Co-authored-by: Phil Henderson <phillip.henderson@intel.com>