DAOS-11890 vos: Process removals to merge adjacent minor epoch extents
Minor epoch writes can cause errors when inserting removal records.

Modify the algorithm when finding records to remove to merge overlapping
extents with the same minor epoch to avoid -DER_NO_PERM.

Replace the hack here with an assertion.  Fix the evtree test so it
fails when this happens (note, it would have failed if that fix were
in place previously).

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>