DAOS-9206 dfs: add container hints for better object class selection (#9375)
DAOS object class selection today is either done explicitly by the user or otherwise automatically depending on the type of the object. However this has proven to be still not sufficient from a user perspective as there is not control on the sharding with automatic selection. For example:

EC is automatically chosen for files / arrays, but would be be bad for small files vs replication
S1 or G1 is chosen for directories, but for fat directories, SX/GX is better
This PR adds some flags to POSIX containers passed at creation time:

dir_oclass (to set explicit oclass for directories to distinguish from files)
hints: single/max to help DAOS better choose an object class for files or directories

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>