DAOS-11911 test: Fix launch.py error reporting
If launch.py experienced an error before it could establish a file
handler for logging debug messages to a file, it could fail without
providing the reason for the failure.  This could happen when running
manually if the python avocado libraries where not installed or
misconfigured. The console logger will now report all messages until the
file handler is configured, at which point the console logger will be
updated to only report info messages.

Required-githooks: true

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>