DAOS-11915 ddb: Reregistration of btree doesn't fail
It was discovered by the ddb tool that when vos_self_init/fini are
called repeatedly, the container btree class is never really
'unregistered' so registering again causes it to fail with
-DER_EXIST. This changes that behavior so that if registering and
the class already exists, it will check if it's registering the
same thing, then returns success. The registration process does
not allocate memory or resources so there is no harm in multiple
registrations.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>