DAOS-9799 chk: Allow policies to be set when starting checker
When running dmg check start, there is a new -p,--policies
argument that accepts a list of class:action policies to
tell the checker how to handle inconsistencies. These policies
are optional, as the default action will be chosen if the
request does not specify an action.

Example:
  * dmg check start -p CIC_POOL_BAD_LABEL:CIA_TRUST_PS,CIC_POOL_BAD_SVCL:CIA_INTERACT

This will tell the checker to trust the pool service label instead
of the MS label in the case of a label discrepancy, in the case of
a bad pool service ranks list, will prompt the administrator to make
a resolution choice via dmg check query/dmg check repair.

Additionally, two new commands are added to manage persistent
policy settings (stored in the MS DB):

  * dmg check get-policy [Class,Class,etc]
  * dmg check set-policy Class:Action,Class:Action,etc.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>