DAOS-12010 bio: Replace dRPC resets LED on non-faulty device
After a daos_server restart, an SMD entry may report device state which
is out of sync with LED state.

In order to rectify this situation, enable the device replace dRPC
handler to reset the LED for an (old/new) UUID that is not in a faulty
state. Then LED state can be reset for a non-faulty device.

Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>