DAOS-11951 test: Use 5 access points for pool/svc.py
The number of ranks we can stop depends on the number
of MS replicas, or access points. For this 5-server
test, we need 5 access points to be able to stop 2
ranks.

Currently, the test uses 3 access points when there
are 3 or more servers. Update test.py so that we can
specify the number of access points from test yaml.

Revert the change in PR10496 regarding the PS
replicas as that's irrelevant.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: test_pool_svc
Required-githooks: true

Signed-off-by: Makito Kano <makito.kano@intel.com>