DAOS-3916 pool: add filter containers API
With this change a new libdaos API is added for a caller to
request a list of a pool's containers that meet a specified
metadata filtering criteria. A new CaRT RPC POOL_FILTER_CONT
is added to contact the pool/container service that executes
the filtering and returns information for the matching containers.

A use case envisioned for the interface is in a tool to identify
what containers are least recently used, e.g., to decide what
to migrate away from the daos storage tier. For example, containers
whose "latest metadata open time is more than one month ago".

New datatypes are added to the libdaos API:
- daos_pool_cont_filter_t:
    Used by callers to input container matching criteria to the
    daos_pool_filter_cont() function, determining what containers to
    return. A filter consists of one or more "parts" each defining a
    container metadata key / value comparison to test. The individual
    part matching outcomes are combined with logical AND or logical OR
    to determine if a container matches the overall filter criteria.
- daos_pool_cont_filter_part_t:
    a filter part defines a boolean condition consisting of a
    metadata key, comparison operator, and value (of type uint64_t
    or string). The container metadata key's current value is compared
    with the specified value (e.g., with ==, !=, <, <=, >, >=).
    This patch implements only uint64_t typed value comparisons.
- daos_pool_cont_info2:
  an extended structure type containing more container information
  than daos_pool_cont_info returned by daos_pool_list_containers().
  For example, the new type include daos_cont_info_t that would be
  returned if a subsequent daos_cont_query() call were performed.
  Placeholder pointers are left in the structure for potential
  future expansion of the API (e.g., to return container properties).

New public signatures and API implementation is added:
- daos_pool_cont_filter_init() to initialize an empty filter
- daos_pool_cont_filter_add() to dynamically add a part to a filter
- daos_pool_cont_filter_fini() finalize a filter and free its parts
- daos_pool_filter_cont() to run the filter (via RPC to pool service).

The libdaos client infrastructure, dc_funcs[] array has a new entry
for this API (dc_pool_filter_cont), and also inserts reserved
placeholders for future API additions.

Two new tests are added to daos_test -p (pool tests) to exercise
the new APIs. The tests invoke a new test function
filter_containers_test() with either no containers in the pool,
or many containers (e.g., 16). The test function invokes the API to
match none, all, or some containers in the pool, using single or
two-part filters and exercising a diversity of keys and comparison
operators. It also invokes the API with conts == NULL (to return
the number of containers), or with an allocation of info structures
to retrieve/verify the idenfitying information of the matches.

The engine container service code is refactored to read metadata
needed to fill a daos_cont_info_t structure, with a common
function cont_info_read() called from container query and from
the pool service when performing this container filtering.

The container service provdes a ds_cont_filter() API for the
pool service to invoke (specifying a filter, getting an array
of matching container information on successful return). The
implementation iterates through all containers in the service,
performing individual filter part and overall filter matching
on each container.

Features: pool

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>