DAOS-11605 common: fix to calculate group address
group header is stored in the first 32k of arena, the first
bit of bitmap is reserved for this, group address should
not add header size, otherwise, last group will be out of boundary.

extend adt_no_space_1 which help reproduce this, also it is
a test for free memory management testing which will be pushed
later.

Required-githooks: true
Signed-off-by: Wang Shilong <shilong.wang@intel.com>