DAOS-11887 test: Remove control/dmg_system_leader_query
This test was written before multiple access points was
supported, and would need to be updated in order to avoid
intermittent failures. Additionally, the test is redundant
in that the leader query/discovery mechanisms are tested
constantly as part of the normal course of running tests.

Features: control
Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>