DAOS-623 bio: Workaround compilation failure
I'm not sure we should be passing umm to bio_yield. Perhaps it's
better handled with a macro for use in places we may be in a
transaction.
Anyway, rather than pass a umm to the vio_replay routine, just pass
and handle NULL parameter to bio_yield.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>