DAOS-3916 container: add number of handles to rdb, query output
Before this change, the daos_cont_info_t structure returned by
daos_cont_open() and daos_cont_query() includes an unnecessary
value, ci_redun_fac (it is otherwise available in the properties).
Additionally, callers to these APIs and daos_pool_filter_containers()
have no visibility into the number of open handles on a container.

With this change, the daos_cont_info_t structure replaces the uint32_t
member ci_redun_fac with ci_nhandles. The daos utility output (both
human-readable and JSON format) is updated to output this quantity.
The libdaos API minor version is also updated. The container open and
query CaRT RPC protocols are  updated to return (where previously a pad
value was specified) this 32-bit quantity. The protocol version is
not changed, since it was already updated for this feature
(metadata time tracking, and filtering based on metadata criteria).
The daos_pool_filter_containers() API and implementation is updated
to allow filtering based on a container's number of open handles.
The daos_test pool and container test suites are updated to utilize
the changes in the API functions and returned information.

The daos_engine container open, query and close handling is updated
to get and increment/decrement the number of handles maintained in rdb.
And engine code that performs pool/container metadata layout upgrades
is updated to initialize the number of handles to zero when upgrading
an older container (e.g., one created with DAOS v2.2). The pool filter
containers handling is updated to process filters that include
conditional tests based on the number of container open handles.

Other changes include:
- a small fix to pool_cont_filter_is_valid() engine code to inspect the
  key and comparison function specified in the RPC against the proper
  minimum and maximum bounds.
- a small change to the pool/container upgrade code for metadata times,
  updating modify time only (not open + modify times) for containers
  that already have metadata times in their rdb. This will fix a future
  upgrade scenario (from DAOS v2.4 to the following version).

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>