DAOS-12015 control,bio: Add timeout option to LED identify command (#10768)
VMD-enabled SSDs can be identified by quickly blinking an LED on the
device using the command dmg storage led identify. This change adds the
ability to specify a timeout option on the commandline, after which the
SSD LED will stop blinking. If no --timeout value is specified on the
commandline, the SSD LED will blink indefinitely. As timeout values are
tied independently to each SSD, different SSDs can be set to a blinking
state for different lengths of time if multiple command invocations use
different values for different SSDs.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>