DAOS-12145 chk: misc fixes for chk related issues - I
It contains the following fixes:

1. Wait pool map refresh before next step check.

If the check engine refreshes the pool map, then it needs to wait
for a while until the pool map has been dispatched to all related
pool shards, that will avoid group versioning related trouble for
subsequent container based check.

2. Auto re-start check from the beginning if former check done.

That will avoid explicitly specifying '-r' option when start check
every time after the first check completed.

3. Allow to start pool service with lost some pool shards.

That is useful when the storage for some pool shards are broken
(such as related vos-x file under the pool directory is lost).

4. Use the same sequence# if report an inconsistency multiple times

For an inconsistency, its sequence# is assigned when it is reported
for the first time. If the first report is for ask for interaction,
then the next report after the repair should reuse the sequence# to
allow control plane to locate the original interaction report.

5. Dedicated ULT on sys SX to handle check engine rejoin.

Signed-off-by: Fan Yong <fan.yong@intel.com>