DAOS-12145: chk: misc fixes for chk related issues - II
It contains the following fixes:

1. Use the same sequence if report an inconsistency multiple times

Continue the fixes in PR#10909, if the first report is for ask for
interaction from non-leader, then the next report after the repair
should reuse the sequence to allow the control plane to locate the
original interaction report.

2. Remove act_msgs from check report upcall

The control plane uses act_details for check query output, that is
more meaningful for the users.

3. Offer pool label and container label for check report upcall

The daos engine will transfer related pool label and the container
label (if applicable) to control plane via check report upcall. It
is more meaningful (than related UUID) for the users.

Signed-off-by: Fan Yong <fan.yong@intel.com>