DAOS-12220 chk: avoid modify const string when upcall
When leader report inconsistency to control plane for interaction
with the user, it may offer some const strings for describing the
inconsistency. The upcall logic should not modify such strings to
avoid DRAM corruption, instead, we will duplicate the strings and
modify the duplicated ones.

On the other hand, the pool ULT that is driven by check main logic
to handle related pool inconsistency should not remove related cpr
from the pool tree to avoid leaking itself in ABT stack.

Signed-off-by: Fan Yong <fan.yong@intel.com>