DAOS-11950 object: Infinite Retry Loop Fix (#10761)
When there is a checksum error on fetch, the request will be marked
as retry and the next replica/shard is fetched. If the next shard is
drained and invalid then it's skipped causing the same previous
shard to be tried again. This patch updates next shard selection
logic to ensure that the shard is valid.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>