DAOS-12121 Build: Backport Unit Test fixes
Skip-bullseye: false

Backport running the Unit Tests and Bullseye on EL-8.

Backport Unit tests no longer using NFS workspace mounts.

The default disabled malware scan code in this branch is not current
and is redundant to the scanning being done in other branches.
It is simpler to remove it instead of updating the malware scan code.

Move Unit Tests to EL-8 as we will not be able to support building SPDK
on EL-7 for use in CI with newer Docker hosts for NLT and Memcheck.

Remove non-functional Coverity stage.

Signed-off-by: John E Malmberg <john.e.malmberg@intel.com>