DAOS-12181 ds3: Don't use stack allocated variables
The function ds3_obj_read() allocates IOV and SGL variables on the stack
for the underlying call to dfs_read(). This pattern works for
synchronous calls, but it causes a segmentation violation for
asynchronous read operations (i.e., when the event structure passed to
dfs_read is not NULL).

Fix this by adding an SG list and IOV structures to the ds3_obj and
using those variables in the call to dfs_read().

Signed-off-by: Chuck Tuffli <chuck.tuffli@hpe.com>