DAOS-11880 engine: Scale size of metrics shmem region (#10947)
Previously, the metrics shared memory region was created with a
static size. However many metrics are now created on a per-target
basis. The size needs to scale with the number of targets.

The estimates of the number of standard metrics and number of
target metrics are higher than the actual number, to allow some
buffer space for future additions.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>