DAOS-12219 control: Add DataVersion field to some responses
In cases where multiple replicas may respond to a given client
request, include the version of the system database in the
response so that the client may choose from the newest in
order to avoid a stale read.

Also updates some unrelated protobuf files to bring them in
line with the output produced by the tools included in el8.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>