DAOS-11517 test: deployment/network_failure.py - Create pool across 4… (#10195) (#11054)
For the network isolation test, we used to create a pool
across two ranks in one node. However, this may cause an
issue with pool service replica when the network interface
is taken down.

Pool service leader can be located in any node regardless
of the pool’s location, so if we happen to take down the
interface on the node where the leader is located, the
pool service will not be accessible and will not be
recovered. Then the subsequent operations may fail.

To resolve this issue, create a pool across four ranks
in two nodes with --nsvc=3, 3 service replicas. Then we
should be able to recover even when the leader is
affected.

Also added the dmg system query check after bringing
down/up the network interface. When we update the
interface, some ranks may get excluded after a while.
i.e., they don't get excluded right after the update. Thus,
wait for 1 min, then repeatedly call dmg system query to
check the status.

Signed-off-by: Makito Kano <makito.kano@intel.com>