DAOS-12168 dfs: add new API to check consistency of DFS namespace (#10961)
DAOS-12168 dfs: add new API to check consistency of DFS namespace

New API to scan through a DFS container and mark all OIDs as seen in an OID table, then scan the OID table to check for unmarked objects that are were not seen in the DFS namespace and print, punch, or add those leaked objects to lost+found.

For the l+f implementation, for now use default chunk size for files. This can cause corruptions if the chunk size was changed.