DAOS-11831 test: Add restart after bringing up network interface (#10565) (#11079)
After bringing up the network interface, some ranks may
become excluded. If that happens, restart the servers
and check again. If they're still not joined, fail the test.

If the server crashes before the 60 sec, we break and move
on to the next step. If it doesn't crash, we wait for
60 sec.

Remove wait_for_ranks_to_join() after bringing down network.

Signed-off-by: Makito Kano <makito.kano@intel.com>