DAOS-12130 test: Cross check core file collection with failed test (#10944)
It’s evident that we can have tests cause segfaults and yet the test passes.
As a safety-net to this situation, it would be useful if the test framework
cross-checked that a test that has produced one or more core files has also
been registered as a FAIL or ERROR.  If it has not, the framework should
log a test failure reflecting that, bringing attention to the fact that a
test is causing a silent segfault.

Signed-off-by: Maureen Jean <maureen.jean@intel.com>