DAOS-12384 control: Support fabric authorization keys
Adds a new server-side config parameter for a fabric auth key
which will be set on the server side and also propagated to
the client applications via the agent.

Bonus: Also adds a new client_env_vars section to the server
config that allows custom client-side environment variables
to be defined. The client environment may override these
variables with locally-defined values.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>