commit 5dba5a18e724c0d6f239b9cfa063bc6a8d2db5f0 Author: Michael MacDonald <mjmac.macdonald@intel.com> Date:   Tue Jan 17 14:01:06 2023 -0500
DAOS-12209 control: Don't clean up Creating pools on MS step-up (#11157)

When a MS leader steps up, it checks the list of pools for any
not found in a Ready state. With this change, a pool in the
Creating state will be ignored rather than destroyed, on the
theory that the create may be completing on the previous leader.

Update the pool svc RAS event handler to automatically set
a Creating pool to Ready, on the theory that the pool create
has most likely completed, given that it is sending us a
svc ranks update event.

Finally, modify dc_mgmt_pool_find() to retry on -DER_NONEXIST
errors which can result from stale MS replica reads.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>