DAOS-12363 test: Add 'storage: auto' test yaml option
Add support for a 'storage: auto' test yaml option that will result in
an extra yaml file being generated and used to define the daos_server
storage configuration.  This storage configuration will use the
available PMEM, NVMe, and/or VMD storage detected on each server host
specified by the test.

This initial implementation will target a few tests to utilize this
feature with additional tests to be added in the future.

Test-tag-vm: pr
Test-tag: test_agent_failure test_mdtest_small test_ior_small test_ior_intercept test_ioconsistency

Required-githooks: true

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>