DAOS-12039 test: Fix NVMe recovery tests
NVMe recovery tests are currently timing out because automatic fault
detection is assumed to be active in the tests but is disabled by
default in the engine.

Fix by moving test specific faulty check to the beginning of
bio_monitor.c:is_bbs_faulty() so true is returned when fail loc has
been set (from daos_nvme_recovery.c) regardless of whether auto
detection is enabled or not.

Test-tag: daos_core_test_nvme
Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>