DAOS-8938 tools: Fix --path handling for daos cont commands
As --path support was bolted on during the transition from
--pool/--cont flags to positional arguments, some regressions
crept in. This commit refactors things to centralize logic
and responsibility for consistent application of the argument
handling rules.

Also adds positional arg support for the (get|set)-(attrs|props)
subcommands for consistency with the dmg command surface.

Features: daos_cmd
Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>