DAOS-12124 control: Fix pool create error after format
When the --size=% flag is used, the pool create command
performs a storage scan in order to determine the actual
available storage across the ranks. In cases where the
system has not started yet (i.e. pool create is run
immediately after storage format completes), the user may
receive a confusing error message. This patch changes the
flow to perform a system query before storage scan so that
the user receives a normal error if the MS hasn't started
yet.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>