DAOS-12415 test: Allow for tests which may produce core files. (#11326)
Add logic to not fail the DaosBuild test if it creates a forefile,
it sometimes creates a corefile from conftest which is expected
and should not cause a test failure.
Fix a bug where the program that generated the core file was read
incorrectly.
Add some debugging where the DaosBuild test fails.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>