DAOS-12524 control: Return created pool on retry
Since #11157 landed, there is a possibility that a retried
pool create may encounter a pool that was automatically
promoted to the Ready state. Instead of returning -DER_ALREADY
(which is unhandled and unhelpful) to the dmg side, just
populate the pool create response as if the retried create handler
was responsible for doing it.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>