DAOS-12364 gurt: Delete old metrics shmem region on startup
If the user increases the number of targets in the server config,
the size of the old shmem region may not be large enough for
per-target metrics. Instead of reattaching to the old region, we
now delete it if it exists before creating the new one.

Features: telemetry

Required-githooks: true

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>