DAOS-12610 vos: Add iterator API to validate current position
In some cases, we may yield in the middle of a callback and need to
request a full reprobe.  The checksum scrubber is such a case. It could,
in theory, read a value, start processing it for checksum scrubbing,
and have the value deleted before it is able to continue.  If this
happens, we can end up reading bad data, or worse, causing a system
crash.

Features: scrubber

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>