lDAOS-8714 pool: split connect handler into 2 ps_lock sections
Before this change, nearly the entire ds_pool_connect_handler()
execution was done while holding ps_lock as a writer, including
the bulk transfer of the pool map buffer to the client.

In certain test configurations using the ofi tcp provider,
and (in one example) with client using a combination of asynchronous
pool connect, and synchronous pool query (indirectly via statfs/dfuse),
the client can deadlock (here, waiting for pool query reply, meanwhile
on the PS leader engine, the connect handler holds the ps_lock as a
writer, preventing the query handler from running). This occurs with
the tcp provider because crt is not progressed in the background
(in which case 2-sided cooperation on bulk transfers is needed).

With this change, the engine ds_pool_connect_handler() execution is
split into a phase 1 taking ps_lock as a reader, the bulk transfer
of the pool map buffer (under no lock), and phase 2 taking ps_lock
as a writer (during which rdb updates associated with connect occur).

Features: pool

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>