DAOS-12661 ds3: Don't use stack allocated variables for async write
The function ds3_obj_write() allocates IOV and SGL variables on the stack
for the underlying call to dfs_write(). This pattern usually work, but
it can cause a segmenation fault if a different thread is polling the
event queue than the thread issuing the writes.  If thread that polls the
event queue schedules the task before the I/O issuing thread that
initiates the ds3_obj_write(), then the stack allocated variables
for the task might now be corrupted by the I/O issuing thread.

Fix this by allocating  ds3_obj_args_t argument that is then freed via
event completion callback.

Signed-off-by: Steve Lahr <steven.lahr@hpe.com>