DAOS-12308 control: Fix dmg system stop RC
Return code should be non-zero if sys stop fails due to non-responsive
ranks. To fix, ensure the Errored field is set appropriately in a
MemberResult struct each time the NonResponsive state is set. Change
all create calls to use the NewMemberResult() helper which will do
this for us.

- MemberResults with unresponsive state have errored flag set
- RC reflects member state error appropriately
- Updated unit tests

Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>