DAOS-12387 cart: Improve swim rankless operation (#11439)
When a swim member whose self ID is SWIM_ID_INVALID (e.g., before the
engine receives its rank from the MS) receives an incoming swim RPC, it
replies with -DER_UNINIT. When a swim member receives such a -DER_UNINIT
reply, it simulates an ALIVE update to indicate that the sender of the
reply is alive. This turns out to be insufficient if a rankless swim
member becomes suspected. The simulated ALIVE update cannot override the
SUSPECT update, as they have the same incarnation. Consequently, there
will eventually be a DEAD update about the rankless swim member, even
though it is actually alive.

This patch changes the swim RPC handler so that the rankless swim member
will simulate the ALIVE update itself, shorting its local swim state,
instead of replying with -DER_UNINIT. The rankless swim member will
infer its rank (only for handling this one RPC) from the RPC header. If
the incoming RPC contains a suspicion about it, this swim member will
have a chance to bump its incarnation and simulate an ALIVE update that
can override the SUSPECT update.


Required-githooks: true

Signed-off-by: Li Wei <wei.g.li@intel.com>