DAOS-11040 common: Move generic slab registration to pool open/create
Since the changes in slab registration to add common allocation sizes,
we don't need to define them in in VOS but can move them internal to
umem.  This allows us to take advantage of common allocation sizes
that just use the generic API.   We can potentially use this for
the reserve API too but I have left that for another patch.

This should reduce the number of slabs that are allocated overall and
reduce overall allocation overhead.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>