DAOS-12683 csum: RPC Unit Tests (#11498)
Introduction of rpc unit testing. This initial change only has
tests for the encoding, decoding, and freeing of checksum related
structures. It will be easy to add other unit style tests for other
structures.

This change also makes it easier to measure how much time is spent in
the rpc functions. This is especially helpful for decoding when
memory allocation occurs. A future change will add timing measurements
for the checksum related RPC functions.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>