DAOS-12180 dfs: extend fs tool to support fixing entry types
In some scenarios, an entry might get corrupted (mode bits, mtime, ctime, etc.). Most of the these are correctable using regular linux commands (chmod, set mtime/ctime, etc.), but others are not, including a corrupted type bit. This PR extend the daos fs tool to add a fix-type option would allows a user to fix the type. The fix-chunk-size command was removed in favor of this new option that allows also to fix-entry and specific chunk-size as something to fix. This option can be extended further for other entry options.

Required-githooks: true

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>